requirements.txt
# requirements.txt

# Core AI Framework (includes evals)
pydantic-ai[logfire]

# Anthropic (your API provider)
anthropic>=0.25.0

# RAG / Vector Store
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Observability
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
logfire

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
pydantic-settings>=2.3.0

# Database (local memory storage)
tinydb>=4.8.0

# Text processing
tiktoken>=0.7.0

gradio>=4.0.0

mcp[cli]>=1.0.0